# PERF: improve reading of geoarrow encoded Parquet (avoid converting coords to geopandas object dtype)