@jorisvandenbossche 2024-06-07
Opened https://github.com/apache/arrow/issues/42026 for the pyarrow perf issue